+++ topobank/analysis/tests/test_functions.py
@@ -16,6 +16,7 @@ from topobank.analysis.functions import (
 
 from topobank.manager.tests.utils import SurfaceFactory, TopographyFactory
 
+
 ###############################################################################
 # Helpers for doing tests
 ###############################################################################
@@ -39,13 +40,12 @@ class FakeTopographyModel:
 
 
 def test_height_distribution_simple_line_scan():
-
-    x = np.array((1,2,3))
-    y = 2*x
+    x = np.array((1, 2, 3))
+    y = 2 * x
 
     info = dict(unit='nm')
 
-    t = NonuniformLineScan(x,y,info=info).detrend(detrend_mode='center')
+    t = NonuniformLineScan(x, y, info=info).detrend(detrend_mode='center')
 
     topography = FakeTopographyModel(t)
 
@@ -55,8 +55,8 @@ def test_height_distribution_simple_line_scan():
 
     assert result['name'] == 'Height distribution'
     assert result['scalars'] == {
-            'Mean Height': {'value': 0,               'unit': 'nm'},
-            'RMS Height':  {'value': math.sqrt(4./3), 'unit': 'nm'},
+        'Mean Height': {'value': 0, 'unit': 'nm'},
+        'RMS Height': {'value': math.sqrt(4. / 3), 'unit': 'nm'},
     }
 
     assert result['xlabel'] == 'Height'
@@ -66,8 +66,8 @@ def test_height_distribution_simple_line_scan():
 
     assert len(result['series']) == 2
 
-    exp_bins = np.array([-1, 1]) # expected values for height bins
-    exp_height_dist_values = [1./6,2./6] # expected values
+    exp_bins = np.array([-1, 1])  # expected values for height bins
+    exp_height_dist_values = [1. / 6, 2. / 6]  # expected values
     series0 = result['series'][0]
     np.testing.assert_almost_equal(series0['x'], exp_bins)
     np.testing.assert_almost_equal(series0['y'], exp_height_dist_values)
@@ -77,9 +77,8 @@ def test_height_distribution_simple_line_scan():
 
 
 def test_slope_distribution_simple_line_scan():
-
-    x = np.array((1,2,3,4))
-    y = -2*x
+    x = np.array((1, 2, 3, 4))
+    y = -2 * x
 
     t = NonuniformLineScan(x, y).detrend(detrend_mode='center')
 
@@ -91,8 +90,8 @@ def test_slope_distribution_simple_line_scan():
 
     assert result['name'] == 'Slope distribution'
     assert result['scalars'] == {
-            'Mean Slope (x direction)': dict(value=-2., unit='1'),  # absolute value of slope
-            'RMS Slope (x direction)':  dict(value=2., unit='1'),   # absolute value of slope
+        'Mean Slope (x direction)': dict(value=-2., unit='1'),  # absolute value of slope
+        'RMS Slope (x direction)': dict(value=2., unit='1'),  # absolute value of slope
     }
 
     assert result['xlabel'] == 'Slope'
@@ -102,8 +101,8 @@ def test_slope_distribution_simple_line_scan():
 
     assert len(result['series']) == 2
 
-    exp_bins = np.array([-2.33333333333, -2, -1.66666666666])  # for slopes
-    exp_slope_dist_values = [0, 3, 0]  # integral with dx=1/3 results to 1
+    exp_bins = np.array([-2-1/1500, -2, -2+1/1500])  # for slopes
+    exp_slope_dist_values = [0, 1500, 0]  # integral with dx=1/3 results to 1
     series0 = result['series'][0]
     np.testing.assert_almost_equal(series0['x'], exp_bins)
     np.testing.assert_almost_equal(series0['y'], exp_slope_dist_values)
@@ -113,15 +112,14 @@ def test_slope_distribution_simple_line_scan():
 
 
 def test_curvature_distribution_simple_line_scan():
-
     unit = 'nm'
     x = np.arange(10)
-    y = -2*x**2 # constant curvature
+    y = -2 * x ** 2  # constant curvature
 
     t = NonuniformLineScan(x, y, info=dict(unit=unit)).detrend(detrend_mode='center')
     topography = FakeTopographyModel(t)
 
-    bins = np.array((-4.75,-4.25,-3.75,-3.25)) # special for this test in order to know results
+    bins = np.array((-4.75, -4.25, -3.75, -3.25))  # special for this test in order to know results
     result = curvature_distribution(topography, bins=bins)
 
     assert sorted(result.keys()) == sorted(['name', 'scalars', 'xlabel', 'ylabel', 'xunit', 'yunit', 'series'])
@@ -140,7 +138,7 @@ def test_curvature_distribution_simple_line_scan():
 
     assert len(result['series']) == 2
 
-    exp_bins = (bins[1:]+bins[:-1])/2
+    exp_bins = (bins[1:] + bins[:-1]) / 2
     exp_curv_dist_values = [0, 2, 0]
 
     # integral over dx= should be 1
@@ -155,10 +153,9 @@ def test_curvature_distribution_simple_line_scan():
 
 
 def test_power_spectrum_simple_nonuniform_linescan():
-
     unit = 'nm'
     x = np.arange(10)
-    y = -2*x**2 # constant curvature
+    y = -2 * x ** 2  # constant curvature
 
     t = NonuniformLineScan(x, y, info=dict(unit=unit)).detrend(detrend_mode='center')
     topography = FakeTopographyModel(t)
@@ -184,9 +181,8 @@ def test_power_spectrum_simple_nonuniform_linescan():
 
 
 def test_autocorrelation_simple_nonuniform_topography():
-
     x = np.arange(5)
-    h = 2*x
+    h = 2 * x
 
     info = dict(unit='nm')
 
@@ -203,7 +199,6 @@ def test_autocorrelation_simple_nonuniform_topography():
 
 
 def test_variable_bandwidth_simple_nonuniform_linescan():
-
     x = np.arange(5)
     h = 2 * x
     info = dict(unit='nm')
@@ -225,24 +220,20 @@ def test_variable_bandwidth_simple_nonuniform_linescan():
 
 
 @pytest.fixture
-def simple_2d_topography():
+def simple_linear_2d_topography():
+    """Simple 2D topography, which is linear in y"""
     unit = 'nm'
     info = dict(unit=unit)
-
     y = np.arange(10).reshape((1, -1))
     x = np.arange(5).reshape((-1, 1))
-
     arr = -2 * y + 0 * x  # only slope in y direction
-
-    t = Topography(arr, (10, 5), info=info).detrend('center')
-
+    t = Topography(arr, (5, 10), info=info).detrend('center')
     return t
 
 
-def test_height_distribution_simple_2d_topography(simple_2d_topography):
-
-    exp_unit = simple_2d_topography.info['unit']
-    topography = FakeTopographyModel(simple_2d_topography)
+def test_height_distribution_simple_2d_topography(simple_linear_2d_topography):
+    exp_unit = simple_linear_2d_topography.info['unit']
+    topography = FakeTopographyModel(simple_linear_2d_topography)
     result = height_distribution(topography, bins=10)
 
     assert sorted(result.keys()) == sorted(['name', 'scalars', 'xlabel', 'ylabel', 'xunit', 'yunit', 'series'])
@@ -261,8 +252,8 @@ def test_height_distribution_simple_2d_topography(simple_2d_topography):
 
     assert len(result['series']) == 2
 
-    exp_bins = np.array([-8.1, -6.3, -4.5, -2.7, -0.9,  0.9,  2.7,  4.5,  6.3,  8.1])  # for heights
-    exp_height_dist_values = np.ones((10,))*1/(10*1.8)  # each interval has width of 1.8, 10 intervals
+    exp_bins = np.array([-8.1, -6.3, -4.5, -2.7, -0.9, 0.9, 2.7, 4.5, 6.3, 8.1])  # for heights
+    exp_height_dist_values = np.ones((10,)) * 1 / (10 * 1.8)  # each interval has width of 1.8, 10 intervals
     series0 = result['series'][0]
 
     assert series0['name'] == 'Height distribution'
@@ -274,10 +265,9 @@ def test_height_distribution_simple_2d_topography(simple_2d_topography):
     # proposal: use a well tested function instead of own formula
 
 
-def test_slope_distribution_simple_2d_topography(simple_2d_topography):
-
+def test_slope_distribution_simple_2d_topography(simple_linear_2d_topography):
     # resulting heights follow this function: h(x,y)=-4y+9
-    topography = FakeTopographyModel(simple_2d_topography)
+    topography = FakeTopographyModel(simple_linear_2d_topography)
     result = slope_distribution(topography, bins=3)
 
     assert sorted(result.keys()) == sorted(['name', 'scalars', 'xlabel', 'ylabel', 'xunit', 'yunit', 'series'])
@@ -285,9 +275,9 @@ def test_slope_distribution_simple_2d_topography(simple_2d_topography):
     assert result['name'] == 'Slope distribution'
 
     assert pytest.approx(result['scalars']['Mean Slope (x direction)']['value']) == 0.
-    assert pytest.approx(result['scalars']['Mean Slope (y direction)']['value']) == -4.
+    assert pytest.approx(result['scalars']['Mean Slope (y direction)']['value']) == -2.
     assert pytest.approx(result['scalars']['RMS Slope (x direction)']['value']) == 0.
-    assert pytest.approx(result['scalars']['RMS Slope (y direction)']['value']) == 4.
+    assert pytest.approx(result['scalars']['RMS Slope (y direction)']['value']) == 2.
 
     for kind, dir in zip(['Mean', 'RMS'], ['x', 'y']):
         assert result['scalars'][f'{kind} Slope ({dir} direction)']['unit'] == '1'
@@ -299,8 +289,8 @@ def test_slope_distribution_simple_2d_topography(simple_2d_topography):
 
     assert len(result['series']) == 4
 
-    exp_bins_x = np.array([-1./3, 0, 1./3]) # for slopes
-    exp_slope_dist_values_x = [0, 3, 0]
+    exp_bins_x = np.array([-1. / 1500, 0, 1. / 1500])  # for slopes
+    exp_slope_dist_values_x = [0, 1500, 0]
     series0 = result['series'][0]
 
     assert series0['name'] == 'Slope distribution (x direction)'
@@ -308,8 +298,8 @@ def test_slope_distribution_simple_2d_topography(simple_2d_topography):
     np.testing.assert_almost_equal(series0['x'], exp_bins_x)
     np.testing.assert_almost_equal(series0['y'], exp_slope_dist_values_x)
 
-    exp_bins_y = np.array([-4-1. / 3, -4, -4+1. / 3])  # for slopes
-    exp_slope_dist_values_y = [0, 3, 0]
+    exp_bins_y = np.array([-2 - 1. / 1500, -2, -2 + 1. / 1500])  # for slopes
+    exp_slope_dist_values_y = [0, 1500, 0]
     series2 = result['series'][2]
 
     assert series2['name'] == 'Slope distribution (y direction)'
@@ -321,12 +311,11 @@ def test_slope_distribution_simple_2d_topography(simple_2d_topography):
     # proposal: use a well tested function instead of own formula
 
 
-def test_curvature_distribution_simple_2d_topography(simple_2d_topography):
-
-    unit = simple_2d_topography.info['unit']
+def test_curvature_distribution_simple_2d_topography(simple_linear_2d_topography):
+    unit = simple_linear_2d_topography.info['unit']
     # resulting heights follow this function: h(x,y)=-4y+9
 
-    topography = FakeTopographyModel(simple_2d_topography)
+    topography = FakeTopographyModel(simple_linear_2d_topography)
     result = curvature_distribution(topography, bins=3)
 
     assert sorted(result.keys()) == sorted(['name', 'scalars', 'xlabel', 'ylabel', 'xunit', 'yunit', 'series'])
@@ -347,8 +336,8 @@ def test_curvature_distribution_simple_2d_topography(simple_2d_topography):
 
     s0, s1 = result['series']
 
-    exp_bins = np.array([-1./3, 0, 1./3]) # for curvatures
-    exp_curvature_dist_values = [0, 3, 0]
+    exp_bins = np.array([-1. / 1500, 0, 1. / 1500])  # for curvatures
+    exp_curvature_dist_values = [0, 1500, 0]
 
     assert s0['name'] == 'Curvature distribution'
 
@@ -358,16 +347,16 @@ def test_curvature_distribution_simple_2d_topography(simple_2d_topography):
     assert s1['name'] == 'RMS curvature'
     # Not testing gaussian here
 
-def test_curvature_distribution_simple_2d_topography_periodic():
 
+def test_curvature_distribution_simple_2d_topography_periodic():
     unit = 'nm'
     info = dict(unit=unit)
 
     y = np.arange(100).reshape((1, -1))
 
-    arr = np.sin(y/2/np.pi) # only slope in y direction, second derivative is -sin
+    arr = np.sin(y / 2 / np.pi)  # only slope in y direction, second derivative is -sin
 
-    t = Topography(arr, (100,100), periodic=True, info=info).detrend('center')
+    t = Topography(arr, (100, 100), periodic=True, info=info).detrend('center')
     # resulting heights follow this function: h(x,y)=-4y+9
 
     topography = FakeTopographyModel(t)
@@ -381,12 +370,11 @@ def test_curvature_distribution_simple_2d_topography_periodic():
     assert result['scalars']['Mean Curvature']['unit'] == '{}⁻¹'.format(unit)
 
 
-def test_power_spectrum_simple_2d_topography(simple_2d_topography):
-
-    unit = simple_2d_topography.info['unit']
+def test_power_spectrum_simple_2d_topography(simple_linear_2d_topography):
+    unit = simple_linear_2d_topography.info['unit']
     # resulting heights follow this function: h(x,y)=-4y+9
 
-    topography = FakeTopographyModel(simple_2d_topography)
+    topography = FakeTopographyModel(simple_linear_2d_topography)
     result = power_spectrum(topography)
 
     assert sorted(result.keys()) == sorted(['name', 'xlabel', 'ylabel', 'xunit', 'yunit', 'xscale', 'yscale', 'series'])
@@ -409,10 +397,9 @@ def test_power_spectrum_simple_2d_topography(simple_2d_topography):
     # TODO Also check values here as integration test?
 
 
-def test_autocorrelation_simple_2d_topography(simple_2d_topography):
-
+def test_autocorrelation_simple_2d_topography(simple_linear_2d_topography):
     # resulting heights follow this function: h(x,y)=-4y+9
-    topography = FakeTopographyModel(simple_2d_topography)
+    topography = FakeTopographyModel(simple_linear_2d_topography)
     result = autocorrelation(topography)
 
     assert sorted(result.keys()) == sorted(['name', 'xlabel', 'ylabel', 'xscale', 'yscale', 'xunit', 'yunit', 'series'])
@@ -422,9 +409,8 @@ def test_autocorrelation_simple_2d_topography(simple_2d_topography):
     # TODO Check result values for autocorrelation
 
 
-def test_variable_bandwidth_simple_2d_topography(simple_2d_topography):
-
-    topography = FakeTopographyModel(simple_2d_topography)
+def test_variable_bandwidth_simple_2d_topography(simple_linear_2d_topography):
+    topography = FakeTopographyModel(simple_linear_2d_topography)
     result = variable_bandwidth(topography)
 
     assert sorted(result.keys()) == sorted(['name', 'xlabel', 'ylabel', 'xscale', 'yscale', 'xunit', 'yunit', 'series'])
@@ -434,25 +420,23 @@ def test_variable_bandwidth_simple_2d_topography(simple_2d_topography):
 
 
 def test_contact_mechanics_incompatible_topography():
-
     x = np.arange(10)
-    arr = 2*x
+    arr = 2 * x
     info = dict(unit='nm')
-    t = NonuniformLineScan(x,arr, info=info).detrend("center")
+    t = NonuniformLineScan(x, arr, info=info).detrend("center")
     topography = FakeTopographyModel(t)
 
     with pytest.raises(IncompatibleTopographyException):
         contact_mechanics(topography)
 
 
-def test_contact_mechanics_whether_given_pressures_in_result(simple_2d_topography):
-
+def test_contact_mechanics_whether_given_pressures_in_result(simple_linear_2d_topography):
     class ProgressRecorder:
         def set_progress(self, a, nsteps):
             pass  # dummy
 
     given_pressures = [2e-3, 1e-2]
-    topography = FakeTopographyModel(simple_2d_topography)
+    topography = FakeTopographyModel(simple_linear_2d_topography)
     result = contact_mechanics(topography,
                                nsteps=None, pressures=given_pressures, storage_prefix='test/',
                                progress_recorder=ProgressRecorder())
@@ -462,7 +446,6 @@ def test_contact_mechanics_whether_given_pressures_in_result(simple_2d_topograph
 
 @pytest.mark.parametrize('periodic', [True, False])
 def test_contact_mechanics_effective_kwargs_in_result(periodic):
-
     y = np.arange(10).reshape((1, -1))
     x = np.arange(5).reshape((-1, 1))
 
@@ -480,7 +463,7 @@ def test_contact_mechanics_effective_kwargs_in_result(periodic):
                                progress_recorder=ProgressRecorder())
 
     exp_effective_kwargs = dict(
-        substrate_str=('' if periodic else 'non')+'periodic',
+        substrate_str=('' if periodic else 'non') + 'periodic',
         nsteps=10,
         pressures=None,
         hardness=None,
@@ -489,39 +472,32 @@ def test_contact_mechanics_effective_kwargs_in_result(periodic):
     assert result['effective_kwargs'] == exp_effective_kwargs
 
 
-def test_rms_values(simple_2d_topography):
-
-    unit = simple_2d_topography.info['unit']
+def test_rms_values(simple_linear_2d_topography):
+    unit = simple_linear_2d_topography.info['unit']
     inverse_unit = '{}⁻¹'.format(unit)
-    topography = FakeTopographyModel(simple_2d_topography)
+    topography = FakeTopographyModel(simple_linear_2d_topography)
     result = rms_values(topography)
 
-    assert result == [
-        {
-            'quantity': 'RMS Height',
-            'direction': None,
-            'value': np.sqrt(33),
-            'unit': unit,
-        },
-        {
-            'quantity': 'RMS Curvature',
-            'direction': None,
-            'value': 0,
-            'unit': inverse_unit,
-        },
-        {
-            'quantity': 'RMS Slope',
-            'direction': 'x',
-            'value': 0,
-            'unit': 1,
-        },
-        {
-            'quantity': 'RMS Slope',
-            'direction': 'y',
-            'value': 4,
-            'unit': 1,
-        }
-    ]
+    assert result[0]['quantity'] == 'RMS Height'
+    assert result[0]['direction'] is None
+    assert np.isclose(result[0]['value'], np.sqrt(33))
+    assert result[0]['unit'] == unit
+
+    assert result[1]['quantity'] == 'RMS Curvature'
+    assert result[1]['direction'] is None
+    assert np.isclose(result[1]['value'], 0)
+    assert result[1]['unit'] == inverse_unit
+
+    assert result[2]['quantity'] == 'RMS Slope'
+    assert result[2]['direction'] == 'x'
+    assert np.isclose(result[2]['value'], 0)
+    assert result[2]['unit'] == 1
+
+    assert result[3]['quantity'] == 'RMS Slope'
+    assert result[3]['direction'] == 'y'
+    assert np.isclose(result[3]['value'], 2)
+    assert result[3]['unit'] == 1
+
 
 ###############################################################################
 # Testing analysis functions for surfaces
@@ -537,7 +513,7 @@ def _expected(xscale):
 
     expected_y = np.piecewise(expected_x,
                               [expected_x < 1, (expected_x >= 1) & (expected_x <= 5), expected_x > 5],
-                              [lambda x:2*x, lambda x: 3*x/2, lambda x:x])
+                              [lambda x: 2 * x, lambda x: 3 * x / 2, lambda x: x])
     expected_std_err_y = np.piecewise(
         expected_x,
         [expected_x < 1, (expected_x >= 1) & (expected_x <= 5), expected_x > 5],
@@ -619,8 +595,9 @@ def test_psd_for_surface(mocker):
         series=[
             {
                 'name': '1D',  # taken from y=2*x
-                'x': 1e-3*np.array([0.1, 1.5, 2.5, 5.0]),   # small numbers because of xunit=nm⁻¹ compared to xunit=µm⁻¹
-                'y': 1e9*np.array([0.2, 3, 5, 10]),  # very small numbers because nm³->µm³
+                'x': 1e-3 * np.array([0.1, 1.5, 2.5, 5.0]),
+                # small numbers because of xunit=nm⁻¹ compared to xunit=µm⁻¹
+                'y': 1e9 * np.array([0.2, 3, 5, 10]),  # very small numbers because nm³->µm³
             }  # the numbers are scaled here in order to match the units of first topography + reuse known results
         ]
     )
@@ -698,8 +675,8 @@ def test_autocorrelation_for_surface(mocker):
         series=[
             {
                 'name': '1D',  # taken from y=2*x
-                'x': 1e3*np.array([0.1, 1.5, 2.5, 5.0]),   # large numbers because of xunit=nm compared to xunit=µm
-                'y': 1e6*np.array([0.2, 3, 5, 10]),  # larger numbers because nm²->µm²
+                'x': 1e3 * np.array([0.1, 1.5, 2.5, 5.0]),  # large numbers because of xunit=nm compared to xunit=µm
+                'y': 1e6 * np.array([0.2, 3, 5, 10]),  # larger numbers because nm²->µm²
             }  # the numbers are scaled here in order to match the units of first topography + reuse known results
         ]
     )
@@ -777,8 +754,8 @@ def test_variable_bandwidth_for_surface(mocker):
         series=[
             {
                 'name': 'VBM',  # taken from y=2*x
-                'x': 1e3*np.array([0.1, 1.5, 2.5, 5.0]),   # large numbers because of xunit=nm compared to xunit=µm
-                'y': 1e3*np.array([0.2, 3, 5, 10]),        # large numbers because of yunit=nm compared to yunit=µm
+                'x': 1e3 * np.array([0.1, 1.5, 2.5, 5.0]),  # large numbers because of xunit=nm compared to xunit=µm
+                'y': 1e3 * np.array([0.2, 3, 5, 10]),  # large numbers because of yunit=nm compared to yunit=µm
             }  # the numbers are scaled here in order to match the units of first topography + reuse known results
         ]
     )