+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 - added small description on surface landing page (#574) 
 - added small description for bandwidth plot (#576)
 - removed plot title from many plots where not needed (#591)
+- uses SurfaceTopography version 0.92.0 now, 
+  which returns nonuniform PSD only up to a reasonable 
+  wavevector (#568)
 
 ## 0.9.6 (2020-12-04)